+++ lib/controllers/dashboard_controller.dart
@@ -10,7 +10,7 @@ class DashboardController {
   /// Attempts to grab the current user's from Amplify
   static Future<User?> loadSignedIn() async {
     try {
-      var authUser = await Amplify.Auth.getCurrentUser(); // has userId, username
+      var authUser = await Amplify.Auth.getCurrentUser();
 
       var user = User(id: authUser.userId); // TODO: replace this line with a call to the USer API to get the user data
 
@@ -24,6 +24,11 @@ class DashboardController {
       return null;
     }
   }
+
+  // load data
+  // activeTows, completedTows, payoutAmount
+
+  // getTowHistory
 
 
 }