+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 
-## 1.1.0
+## 0.3.0
+* Created DashboardScreen UI elements
+* Add placeholder dashboard_controller
+
+## 0.2.0
 * Updated amplify config
 * Updated onboarding controller to call user/company APIs
 * Updated Company/User models
@@ -8,7 +12,7 @@
 * Create CompanyAPI and UserAPI services
 * Update onboarding screens
 
-## 1.0.0
+## 0.1.0
 * Initialized project structure
 * Created onboarding flow (views, controller, models)
 * Added placeholder screen for dashboard